[vpr] categorize files to rr_graph_type & rr_graph_cost in librrgraph
