--- libs/libarchfpga/src/base_cost_type.h
@@ -1,14 +1,0 @@
-#ifndef BASE_COST_TYPE_H
-#define BASE_COST_TYPE_H
-
-enum e_base_cost_type {
-    DELAY_NORMALIZED,
-    DELAY_NORMALIZED_LENGTH,
-    DELAY_NORMALIZED_FREQUENCY,
-    DELAY_NORMALIZED_LENGTH_FREQUENCY,
-    DELAY_NORMALIZED_LENGTH_BOUNDED,
-    DEMAND_ONLY,
-    DEMAND_ONLY_NORMALIZED_LENGTH
-};
-
-#endif
