--- libs/libarchfpga/src/route_type.h
@@ -1,9 +1,0 @@
-#ifndef ROUTE_TYPE_H
-#define ROUTE_TYPE_H
-
-enum e_route_type {
-    GLOBAL,
-    DETAILED
-};
-
-#endif
