+++ libs/librrgraph/src/base/get_parallel_segs.h
@@ -1,7 +1,7 @@
 #ifndef GET_PARALLEL_SEGS_H
 #define GET_PARALLEL_SEGS_H
 
-#include "unified_to_parallel_seg_index.h"
+#include "rr_graph_type.h"
 #include "physical_types.h"
 
 std::vector<t_segment_inf> get_parallel_segs(const std::vector<t_segment_inf>& segment_inf,
