+++ libs/librrgraph/src/base/rr_graph_utils.h
@@ -1,26 +1,24 @@
 #ifndef RR_GRAPH_UTILS_H
 #define RR_GRAPH_UTILS_H
 
-// Make room in a vector, with amortized O(1) time by using a pow2 growth pattern.
-//
-// This enables potentially random insertion into a vector with amortized O(1)
-// time.
-template<typename T>
-void make_room_in_vector(T* vec, size_t elem_position) {
-    if (elem_position < vec->size()) {
-        return;
-    }
+/* Include header files which include data structures used by
+ * the function declaration
+ */
+#include <vector>
+#include "rr_graph_fwd.h"
+#include "rr_node_types.h"
+#include "rr_graph_view.h"
 
-    size_t capacity = std::max(vec->capacity(), size_t(16));
-    while (elem_position >= capacity) {
-        capacity *= 2;
-    }
+/* Get node-to-node switches in a RRGraph */
+std::vector<RRSwitchId> find_rr_graph_switches(const RRGraph& rr_graph,
+                                               const RRNodeId& from_node,
+                                               const RRNodeId& to_node);
 
-    if (capacity >= vec->capacity()) {
-        vec->reserve(capacity);
-    }
+// This function generates and returns a vector indexed by RRNodeId
+// containing a list of fan-in edges for each node.
+vtr::vector<RRNodeId, std::vector<RREdgeId>> get_fan_in_list(const RRGraphView& rr_graph);
 
-    vec->resize(elem_position + 1);
-}
+int seg_index_of_cblock(const RRGraphView& rr_graph, t_rr_type from_rr_type, int to_node);
+int seg_index_of_sblock(const RRGraphView& rr_graph, int from_node, int to_node);
 
 #endif
