+++ vpr/src/route/rr_graph.cpp
@@ -17,7 +17,7 @@
 #include "vpr_error.h"
 
 #include "globals.h"
-#include "rr_graph_util.h"
+#include "rr_graph_utils.h"
 #include "rr_graph.h"
 #include "rr_graph_area.h"
 #include "rr_graph2.h"
