+++ vpr/src/route/rr_graph2.cpp
@@ -9,7 +9,7 @@
 #include "vpr_error.h"
 
 #include "globals.h"
-#include "rr_graph_util.h"
+#include "rr_graph_utils.h"
 #include "rr_graph2.h"
 #include "rr_graph.h"
 #include "rr_graph_sbox.h"
