+++ vpr/src/route/rr_graph_area.cpp
@@ -10,7 +10,7 @@
 
 #include "globals.h"
 #include "rr_graph.h"
-#include "rr_graph_util.h"
+#include "rr_graph_utils.h"
 #include "rr_graph_area.h"
 
 /* Select which transistor area equation to use. As found by Chiasson's and Betz's FPL 2013 paper
