+++ vpr/src/route/rr_graph_timing_params.cpp
@@ -7,7 +7,7 @@
 
 #include "globals.h"
 #include "rr_graph.h"
-#include "rr_graph_util.h"
+#include "rr_graph_utils.h"
 #include "rr_graph2.h"
 #include "rr_graph_timing_params.h"
